adding support for 64 bit archs

+++ java/com/facebook/crypto/mac/NativeMac.java
@@ -89,7 +89,7 @@ public class NativeMac {
 
   // Used to store the HMAC context.
   @DoNotStrip
-  private int mCtxPtr;
+  private long mCtxPtr;
 
   // The integer value representing failure in JNI world.
   private static native int nativeFailure();
